Move `mask_indices` to `genereate.py` (and move tests too)
