+++ cupy/__init__.py
@@ -395,6 +395,7 @@ from numpy import typename  # NOQA
 from cupy._indexing.generate import c_  # NOQA
 from cupy._indexing.generate import indices  # NOQA
 from cupy._indexing.generate import ix_  # NOQA
+from cupy._indexing.generate import mask_indices  # NOQA
 from cupy._indexing.generate import r_  # NOQA
 from cupy._indexing.generate import ravel_multi_index  # NOQA
 from cupy._indexing.generate import unravel_index  # NOQA
@@ -668,7 +669,6 @@ from cupy._sorting.search import argmax  # NOQA
 from cupy._sorting.search import argmin  # NOQA
 from cupy._sorting.search import argwhere  # NOQA
 from cupy._sorting.search import flatnonzero  # NOQA
-from cupy._sorting.search import mask_indices  # NOQA
 from cupy._sorting.search import nanargmax  # NOQA
 from cupy._sorting.search import nanargmin  # NOQA
 from cupy._sorting.search import nonzero  # NOQA
